requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.3
Flask~=2.1.0
flask_bootstrap==3.3.7.1
flask_login~=0.6.1
flask_sqlalchemy==2.5.1
flask_wtf==1.0.1
SQLAlchemy==1.4.22
Werkzeug~=2.2.1
WTForms==3.0.1
email-validator==1.2.1
APScheduler~=3.9.1
numpy~=1.23.1
Pillow~=9.2.0